requirements.txt
playwright>=1.30.0
SpeechRecognition>=3.8.1
pyttsx3>=2.90
google-generativeai>=0.3.0
python-dotenv>=0.19.2
asyncio>=3.4.3
